+++ Tools/eslint-config-cesium/CHANGES.md
@@ -4,6 +4,7 @@ Change Log
 ### 2.0.0
 
 * Enable [no-floating-decimal](http://eslint.org/docs/rules/no-floating-decimal).
+* Enable [no-use-before-define](http://eslint.org/docs/rules/no-use-before-define).
 * Enable [no-else-return](http://eslint.org/docs/rules/no-else-return).
 * Enable [no-alert](http://eslint.org/docs/rules/no-alert).
 * Enable [no-loop-func](http://eslint.org/docs/rules/no-loop-func).